pref: optimize the diff page selection
